Refactor application structure: move App component to frontend directory, update imports, and remove legacy code.
