+++ src/components/frontend/src/App.jsx
@@ -1,7 +1,7 @@
 // src/App.tsx (lub App.jsx)
 
 import React from 'react'
-import LoginForm from './components/LoginForm'
+import LoginForm from '.src\components'
 
 function App() {
   return (
