+++ src/index.jsx
@@ -1,5 +1,4 @@
-import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './App';
+import App from './components/frontend/src/App';
 
 ReactDOM.render(<App />, document.getElementById('root'));
